only set reminders 10 mins before due
- conserve memory by only actually setting the timers 10 minutes before they are due instead of hours, sometimes days before, wasting memory
- updated reminder limit to 10
- allow for unlimited reminder lengths, instead of 24 day limit
- small optimisations
