+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "reminders",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "Set reminders for things you need to do.",
   "main": "dist/index.js",
   "scripts": {
@@ -18,14 +18,14 @@
   },
   "homepage": "https://github.com/wdhdev/reminders#readme",
   "dependencies": {
-    "@sentry/node": "^8.9.2",
+    "@sentry/node": "^8.10.0",
     "discord.js": "^14.15.3",
     "dotenv": "^16.4.5",
-    "mongoose": "^8.4.1"
+    "mongoose": "^8.4.3"
   },
   "devDependencies": {
-    "@sentry/types": "^8.9.2",
-    "@types/node": "^20.14.2",
-    "typescript": "^5.4.5"
+    "@sentry/types": "^8.10.0",
+    "@types/node": "^20.14.7",
+    "typescript": "^5.5.2"
   }
 }
