+++ src/events/client/ready.ts
@@ -6,6 +6,7 @@ import { exec } from "child_process";
 import globalCommands from "../../scripts/global-commands";
 
 import Reminder from "../../models/Reminder";
+import setReminder from "../../util/setReminder";
 
 const event: Event = {
     name: "ready",
@@ -30,64 +31,24 @@ const event: Event = {
             }, 30 * 1000) // 30 seconds
 
             // Manage timeouts
-            let reminders = await Reminder.find({});
+            const setReminders: string[] = [];
 
-            const dueReminders = reminders.filter(reminder => reminder.due <= Date.now().toString());
+            async function manageExistingTimeouts() {
+                let reminders = await Reminder.find({});
 
-            for(const reminder of dueReminders) {
-                // if(!client.guilds.cache.has(reminder.guild)) {
-                //     await reminder.deleteOne();
-                //     reminders = reminders.filter(r => r !== reminder);
-                //     continue;
-                // }
+                const dueReminders = reminders.filter(reminder => reminder.due <= Date.now().toString());
 
-                await reminder.deleteOne();
-                reminders = reminders.filter(r => r !== reminder);
-
-                const embed = new Discord.EmbedBuilder()
-                    .setColor(client.config_embeds.default)
-                    .setTitle("Overdue Reminder")
-                    .setDescription(reminder.reason)
-                    .addFields (
-                        { name: "Set", value: `<t:${reminder.set.toString().slice(0, -3)}:f>`, inline: true },
-                        { name: "Overdue Since", value: `<t:${reminder.due.toString().slice(0, -3)}:R>`, inline: true }
-                    )
-                    .setFooter({ text: `ID: ${reminder.reminder_id}` })
-                    .setTimestamp()
-
-                try {
-                    const user = client.users.cache.get(reminder.user);
-
-                    await user.send({ embeds: [embed] });
-                } catch {
-                    try {
-                        const channel = client.channels.cache.get(reminder.channel) as Discord.TextChannel;
-
-                        if(!channel) return;
-
-                        await channel.send({ content: `<@${reminder.user}>`, embeds: [embed] });
-                    } catch {}
-                }
-            }
-
-            for(const reminder of reminders) {
-                // if(!client.guilds.cache.has(reminder.guild)) {
-                //     await reminder.deleteOne();
-                //     continue;
-                // }
-
-                const delay = Number(reminder.due) - Date.now();
-
-                client.reminders.set(`${reminder.user}-${reminder.reminder_id}`, setTimeout(async () => {
+                for(const reminder of dueReminders) {
                     await reminder.deleteOne();
-                    client.reminders.delete(`${reminder.user}-${reminder.reminder_id}`);
+                    reminders = reminders.filter(r => r !== reminder);
 
                     const embed = new Discord.EmbedBuilder()
                         .setColor(client.config_embeds.default)
-                        .setTitle("Reminder")
+                        .setTitle("Overdue Reminder")
                         .setDescription(reminder.reason)
                         .addFields (
-                            { name: "Set", value: `<t:${reminder.set.toString().slice(0, -3)}:f>` }
+                            { name: "Set", value: `<t:${reminder.set.toString().slice(0, -3)}:f>`, inline: true },
+                            { name: "Overdue Since", value: `<t:${reminder.due.toString().slice(0, -3)}:R>`, inline: true }
                         )
                         .setFooter({ text: `ID: ${reminder.reminder_id}` })
                         .setTimestamp()
@@ -105,8 +66,30 @@ const event: Event = {
                             await channel.send({ content: `<@${reminder.user}>`, embeds: [embed] });
                         } catch {}
                     }
-                }, delay))
+                }
+
+                for(const reminder of reminders) {
+                    const result = await setReminder(reminder, client);
+
+                    if(result) setReminders.push(`${reminder.user}-${reminder.reminder_id}`);
+                }
             }
+
+            manageExistingTimeouts().then(async () => {
+                setInterval(async () => {
+                    const reminders = await Reminder.find({});
+
+                    if(reminders.length === 0) return;
+
+                    for(const reminder of reminders) {
+                        if(setReminders.includes(`${reminder.user}-${reminder.reminder_id}`)) continue;
+
+                        const result = await setReminder(reminder, client);
+
+                        if(result) setReminders.push(`${reminder.user}-${reminder.reminder_id}`);
+                    }
+                }, 60000)
+            })
         } catch(err) {
             client.logError(err);
         }
